+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed a bug causing an error in generation of requests which have an enum or an UntypedNode as body. [microsoft/kiota-dotnet#537](https://github.com/microsoft/kiota-dotnet/issues/537)
 - Removed redundant additional data property definition in TypeScript.
 - Fixed a bug where the value parameter for serialization/deserialization in TypeScript methods would be missing a doc comment.
+- Fixed a bug where serializing a base type in TypeScript would lead to missing properties from the derived type. [#6511](https://github.com/microsoft/kiota/issues/6511)
 
 ## [1.27.0] - 2025-06-11